tests(loader): add loader_spec
