+++ plugin/rocks.lua
@@ -10,34 +10,9 @@ local adapter = require("rocks.adapter")
 log.trace("loading rocks config")
 local config = require("rocks.config.internal")
 
-local function get_luarocks_lua_dir_from_luarocks()
-    local sc = vim.system({ config.luarocks_binary, "--lua-version=5.1", "which", "luarocks.loader" }):wait()
-    local result = sc.stdout and sc.stdout:match(vim.fs.joinpath("(%S+)", "5.1", "luarocks", "loader.lua"))
-    return result
-end
-
 -- Initialize the luarocks loader
 if config.enable_luarocks_loader then
-    local default_luarocks_binary = vim.fs.joinpath(config.rocks_path, "bin", "luarocks")
-    local luarocks_lua_dir = config.luarocks_binary == default_luarocks_binary
-            and vim.fs.joinpath(default_luarocks_binary, "share", "lua")
-        or get_luarocks_lua_dir_from_luarocks()
-    if luarocks_lua_dir then
-        package.path = package.path
-            .. ";"
-            .. table.concat({
-                vim.fs.joinpath(luarocks_lua_dir, "5.1", "?.lua"),
-                vim.fs.joinpath(luarocks_lua_dir, "5.1", "init.lua"),
-            }, ";")
-        vim.env.LUAROCKS_CONFIG = config.luarocks_config
-        local ok, err = pcall(require, "luarocks.loader")
-        if not ok then
-            log.error(err)
-            vim.notify("Failed to initialize luarocks loader: " .. err, vim.log.levels.WARN, {
-                title = "rocks.nvim",
-            })
-        end
-    end
+    require("rocks.loader").enable()
 end
 
 -- Set up the Rocks user command
